+++ src/components/reusable/textInput/textInput.ts
@@ -27,6 +27,7 @@ import errorIcon from '@carbon/icons/es/warning--filled/24';
 export class TextInput extends LitElement {
   static override styles = TextInputScss;
 
+  /** @ignore */
   static override shadowRootOptions = {
     ...LitElement.shadowRootOptions,
     delegatesFocus: true,